Move BasicNode in its own file
